+++ src/commands.ts
@@ -1,0 +1,85 @@
+import * as vscode from 'vscode';
+import { DebugProtocol } from 'vscode-debugprotocol';
+import { URI } from 'vscode-uri';
+import { unescapeAhk } from './util/VariableManager';
+
+interface VariableContextMenuParam {
+  container: DebugProtocol.Scope;
+  variable: DebugProtocol.Variable;
+}
+const removeQuote = (string: string): string => {
+  return string.replace(/(^"|"$)/gu, '');
+};
+const convertToText = (param: VariableContextMenuParam): string => {
+  return unescapeAhk(removeQuote(param.variable.value));
+};
+const convertToNumber = (param: VariableContextMenuParam, radix: number): string => {
+  return Math.ceil(Number(removeQuote(param.variable.value))).toString(radix);
+};
+const convertToHex = (param: VariableContextMenuParam): string => {
+  const number = Math.ceil(Number(removeQuote(param.variable.value)));
+  return 0 < number ? `0x${number.toString(16)}` : `-0x${number.toString(16).replace(/^-/u, '')}`;
+};
+const convertToScientificNotation = (param: VariableContextMenuParam): string => {
+  return Number(removeQuote(param.variable.value)).toExponential();
+};
+const showValue = async(text: string): Promise<void> => {
+  const uri = URI.parse(`valuepreview:${encodeURI(text)}.ahk`);
+  const doc = await vscode.workspace.openTextDocument(uri); // calls back into the provider
+  await vscode.window.showTextDocument(doc, { preview: true });
+};
+export const registerCommands = (context: vscode.ExtensionContext): void => {
+  context.subscriptions.push(vscode.workspace.registerTextDocumentContentProvider('valuepreview', new class implements vscode.TextDocumentContentProvider {
+    public readonly onDidChangeEmitter = new vscode.EventEmitter<vscode.Uri>();
+    public readonly onDidChange = this.onDidChangeEmitter.event;
+    public provideTextDocumentContent(uri: vscode.Uri): string {
+      return decodeURI(uri.path.replace(/\.ahk/u, ''));
+    }
+  }()));
+
+  // View
+  context.subscriptions.push(vscode.commands.registerCommand('vscode-autohotkey-debug.variables-view.viewValue', async(param: VariableContextMenuParam): Promise<void> => {
+    await showValue(param.variable.value);
+  }));
+  context.subscriptions.push(vscode.commands.registerCommand('vscode-autohotkey-debug.variables-view.viewAsText', async(param: VariableContextMenuParam): Promise<void> => {
+    await showValue(convertToText(param));
+  }));
+  context.subscriptions.push(vscode.commands.registerCommand('vscode-autohotkey-debug.variables-view.ViewAsEachBaseNumbers', async(param: VariableContextMenuParam): Promise<void> => {
+    const text = [
+      `[Binary]`,
+      convertToNumber(param, 2),
+      '',
+      `[Decimal]`,
+      convertToNumber(param, 10),
+      '',
+      `[Hex]`,
+      convertToHex(param),
+      '',
+      `[Scientific Notation]`,
+      convertToScientificNotation(param),
+    ].join('\n');
+    await showValue(text);
+  }));
+
+  // Copy
+  context.subscriptions.push(vscode.commands.registerCommand('vscode-autohotkey-debug.variables-view.copyAsText', async(param: VariableContextMenuParam): Promise<void> => {
+    const text = convertToText(param);
+    await vscode.env.clipboard.writeText(text);
+  }));
+  context.subscriptions.push(vscode.commands.registerCommand('vscode-autohotkey-debug.variables-view.copyAsDecimal', async(param: VariableContextMenuParam): Promise<void> => {
+    const decimal = convertToNumber(param, 10);
+    await vscode.env.clipboard.writeText(decimal);
+  }));
+  context.subscriptions.push(vscode.commands.registerCommand('vscode-autohotkey-debug.variables-view.copyAsBinary', async(param: VariableContextMenuParam): Promise<void> => {
+    const binary = convertToNumber(param, 2);
+    await vscode.env.clipboard.writeText(binary);
+  }));
+  context.subscriptions.push(vscode.commands.registerCommand('vscode-autohotkey-debug.variables-view.copyAsHex', async(param: VariableContextMenuParam): Promise<void> => {
+    const hex = convertToHex(param);
+    await vscode.env.clipboard.writeText(hex.startsWith('-') ? `-0x${hex.substr(1)}` : `0x${hex}`);
+  }));
+  context.subscriptions.push(vscode.commands.registerCommand('vscode-autohotkey-debug.variables-view.copyAsScientificNotation', async(param: VariableContextMenuParam): Promise<void> => {
+    const scientificNotation = convertToScientificNotation(param);
+    await vscode.env.clipboard.writeText(scientificNotation);
+  }));
+};